Populate 3d model data
